requirements.txt
fastapi[standard]
langchain_core==0.3.75
langchain_openai==0.3.32
langgraph==0.6.6
langgraph-checkpoint-mongodb===0.2.0
pydantic==2.11.7
python-dotenv==1.1.1
